refactor: вынес вебсокет менеджер в отдельный файл
