+++ backend/app/ws/ws.py
@@ -1,0 +1,24 @@
+from fastapi import WebSocket
+
+class WSManager:
+    def __init__(self):
+        self._clients: set[WebSocket] = set()
+
+    async def connect(self, ws: WebSocket):
+        await ws.accept()
+        self._clients.add(ws)
+
+    def disconnect(self, ws: WebSocket):
+        self._clients.discard(ws)
+
+    async def broadcast(self, message: str):
+        dead: list[WebSocket] = []
+        for ws in self._clients:
+            try:
+                await ws.send_text(message)
+            except Exception:
+                dead.append(ws)
+        for ws in dead:
+            self.disconnect(ws)
+            
+ws_manager = WSManager()
